Add Google Analytics detection
Early version of Google Analytics detection. Checks if GA is present
and if yes, if it uses the AnonymizeIP setting consistently. May have
to add additional URLs to the RegEx pending further data on how GA is
embedded in the wild - there seem to be a large number of endpoints.