Do not detect redirects as third party requests
The domain which the user enters and the domain of the final URL will be
considered trusted (and therefore not third-party). If there is a
redirect in between, it will be considered as third party (e.g.
example.com redirects to eviltracker.com?dest=example.org which
redirects to example.org)