chromedevtools: Rewrite the waiting routine for a page to be ready
Our waiting routine now waits again if there is a pending redirect
issued by either JavaScript or a meta tag. We consider a page to be
stable after there is no scheduled navigation, i.e., an upcoming
redirect. If we are in a stable state, we intercept all upcoming
requests that will change the document so that our page stays stable.
Fixes #17.