dns: Fix MX record host parsing
We always stripped the last character of the host in the dns answer. This is
right for 99% of the cases, since basically everyone uses a dot at the end to
mark a FQDN. There is, however, the possibility to leave out the last dot; in
this case we do not have an FQDN but must add the domain. Moreover, some sites
(example.com, ...) use only a ".", which seems invalid, so we skip those.